Remove `ActiveExecutionService`
